Refactor AI types and move definitions to types.ts
